requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
watchdog>=2.0.0